requirements.txt
#slackclient==1.3.1
slackclient==2.9.3
python-dotenv==0.15.0
beautifulsoup4==4.9.3
lxml==4.6.2
